requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.21
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary